lsm303agr
Struct Lsm303agr

pub struct Lsm303agr<DI, MODE> { /* private fields */ }

LSM303AGR device driver

Implementations

impl<DI, CommE, PinE, MODE> Lsm303agr<DI, MODE>
where DI: ReadData<Error = Error<CommE, PinE>> + WriteData<Error = Error<CommE, PinE>>,

pub fn set_accel_mode_and_odr<D: DelayUs<u32>>( &mut self, delay: &mut D, mode: AccelMode, odr: impl Into<Option<AccelOutputDataRate>>, ) -> Result<(), Error<CommE, PinE>>

Set accelerometer power/resolution mode and output data rate.

Returns Error::InvalidInputData if the mode is incompatible with the given output data rate.

The given delay is used to wait for the sensor to turn on or change modes, according to the times specified in Table 14 and Table 15 in the datasheet. You can opt out of this by using a no-op delay implementation, see embedded_hal_mock::delay::MockNoop for an example of such an implementation.

pub fn get_accel_mode(&mut self) -> AccelMode

Get the accelerometer mode

pub fn set_accel_scale( &mut self, scale: AccelScale, ) -> Result<(), Error<CommE, PinE>>

Set accelerometer scaling factor

This changes the scale at which the acceleration is read. AccelScale::G2 for example can return values between -2g and +2g where g is the gravity of the earth (~9.82 m/s²).

pub fn get_accel_scale(&self) -> AccelScale

Get accelerometer scaling factor

impl<I2C> Lsm303agr<I2cInterface<I2C>, MagOneShot>

pub fn new_with_i2c(i2c: I2C) -> Self

Create new instance of the LSM303AGR device communicating through I2C.

impl<I2C, MODE> Lsm303agr<I2cInterface<I2C>, MODE>

pub fn destroy(self) -> I2C

Destroy driver instance, return I2C bus.

impl<SPI, CSXL, CSMAG> Lsm303agr<SpiInterface<SPI, CSXL, CSMAG>, MagOneShot>

pub fn new_with_spi( spi: SPI, chip_select_accel: CSXL, chip_select_mag: CSMAG, ) -> Self

Create new instance of the LSM303AGR device communicating through SPI.

impl<SPI, CSXL, CSMAG, MODE> Lsm303agr<SpiInterface<SPI, CSXL, CSMAG>, MODE>

pub fn destroy(self) -> (SPI, CSXL, CSMAG)

Destroy driver instance, return SPI bus instance and chip select pin.

impl<DI, CommE, PinE, MODE> Lsm303agr<DI, MODE>
where DI: ReadData<Error = Error<CommE, PinE>> + WriteData<Error = Error<CommE, PinE>>,

pub fn init(&mut self) -> Result<(), Error<CommE, PinE>>

Initialize registers

pub fn acc_set_fifo_mode( &mut self, mode: FifoMode, fth: u8, ) -> Result<(), Error<CommE, PinE>>

Set the accelerometer FIFO mode and full threshold.

The threshold is clamped to [0, 31].

pub fn acc_enable_interrupt( &mut self, interrupt: Interrupt, ) -> Result<(), Error<CommE, PinE>>

Enable accelerometer interrupt.

pub fn acc_disable_interrupt( &mut self, interrupt: Interrupt, ) -> Result<(), Error<CommE, PinE>>

Disable accelerometer interrupt.

pub fn mag_enable_int(&mut self) -> Result<(), Error<CommE, PinE>>

Configure the DRDY pin as a digital output.

pub fn mag_enable_low_pass_filter(&mut self) -> Result<(), Error<CommE, PinE>>

Enable magnetometer low-pass filter.

pub fn mag_disable_low_pass_filter(&mut self) -> Result<(), Error<CommE, PinE>>

Disable magnetometer low-pass filter.

pub fn accel_status(&mut self) -> Result<Status, Error<CommE, PinE>>

Accelerometer status

pub fn acceleration(&mut self) -> Result<Acceleration, Error<CommE, PinE>>

Get measured acceleration.

pub fn mag_status(&mut self) -> Result<Status, Error<CommE, PinE>>

Magnetometer status

pub fn accelerometer_id( &mut self, ) -> Result<AccelerometerId, Error<CommE, PinE>>

Get the accelerometer device ID.

pub fn magnetometer_id(&mut self) -> Result<MagnetometerId, Error<CommE, PinE>>

Get the magnetometer device ID.

pub fn temperature(&mut self) -> Result<Temperature, Error<CommE, PinE>>

Get measured temperature.

pub fn temperature_status( &mut self, ) -> Result<TemperatureStatus, Error<CommE, PinE>>

Temperature sensor status

impl<DI, CommE, PinE> Lsm303agr<DI, MagOneShot>
where DI: ReadData<Error = Error<CommE, PinE>> + WriteData<Error = Error<CommE, PinE>>,

pub fn into_mag_continuous( self, ) -> Result<Lsm303agr<DI, MagContinuous>, ModeChangeError<CommE, PinE, Self>>

Change the magnetometer to continuous measurement mode

impl<DI, CommE, PinE> Lsm303agr<DI, MagContinuous>
where DI: ReadData<Error = Error<CommE, PinE>> + WriteData<Error = Error<CommE, PinE>>,

pub fn into_mag_one_shot( self, ) -> Result<Lsm303agr<DI, MagOneShot>, ModeChangeError<CommE, PinE, Self>>

Change the magnetometer to one-shot mode

After this the magnetometer is in idle mode until a one-shot measurement is started.

impl<DI, CommE, PinE, MODE> Lsm303agr<DI, MODE>
where DI: ReadData<Error = Error<CommE, PinE>> + WriteData<Error = Error<CommE, PinE>>,

pub fn set_mag_mode_and_odr<D: DelayUs<u32>>( &mut self, delay: &mut D, mode: MagMode, odr: MagOutputDataRate, ) -> Result<(), Error<CommE, PinE>>

Set magnetometer power/resolution mode and output data rate.

The given delay is used to wait for the sensor to turn on or change modes, according to the times specified in Table 14 and Table 15 in the datasheet. You can opt out of this by using a no-op delay implementation, see embedded_hal_mock::delay::MockNoop for an example of such an implementation.

pub fn get_mag_mode(&self) -> MagMode

Get magnetometer power/resolution mode.

impl<DI, CommE, PinE> Lsm303agr<DI, MagContinuous>
where DI: ReadData<Error = Error<CommE, PinE>> + WriteData<Error = Error<CommE, PinE>>,

pub fn magnetic_field(&mut self) -> Result<MagneticField, Error<CommE, PinE>>

Get the measured magnetic field.

pub fn enable_mag_offset_cancellation( &mut self, ) -> Result<(), Error<CommE, PinE>>

Enable the magnetometer’s built in offset cancellation.

Offset cancellation is automatically managed by the device in continuous mode.

To later disable offset cancellation, use the disable_mag_offset_cancellation function

pub fn disable_mag_offset_cancellation( &mut self, ) -> Result<(), Error<CommE, PinE>>

Disable the magnetometer’s built in offset cancellation.

impl<DI, CommE, PinE> Lsm303agr<DI, MagOneShot>
where DI: ReadData<Error = Error<CommE, PinE>> + WriteData<Error = Error<CommE, PinE>>,

pub fn magnetic_field(&mut self) -> Result<MagneticField, Error<CommE, PinE>>

Get the measured magnetic field.

pub fn enable_mag_offset_cancellation( &mut self, ) -> Result<(), Error<CommE, PinE>>

Enable the magnetometer’s built in offset cancellation.

Offset cancellation has to be managed by the user in single measurement (OneShot) mode averaging two consecutive measurements Hn and Hn-1.

To later disable offset cancellation, use the disable_mag_offset_cancellation function

pub fn disable_mag_offset_cancellation( &mut self, ) -> Result<(), Error<CommE, PinE>>

Disable the magnetometer’s built in offset cancellation.